refactor(iconbox): separate icon-box from resume, contact list
