+++ src/components/side-bar/contact-list.tsx
@@ -2,6 +2,7 @@ import { HiOutlineMail } from "react-icons/hi";
 import { IoCalendarOutline } from "react-icons/io5";
 import { MdOutlineLocationOn } from "react-icons/md";
 import { TbPhoneCalling } from "react-icons/tb";
+import IconBox from "../icon-box";
 
 import "@/styles/contact-list.css";
 
@@ -9,16 +10,10 @@ import "@/styles/contact-list.css";
 const ContactsList: React.FC = () => {
 	return (
 		<ul className="contacts-list">
-
 			<li className="contact-item">
-
-				<div className="icon-box">
-					<HiOutlineMail />
-				</div>
-
+				<IconBox icon={HiOutlineMail} />
 				<div className="contact-info">
 					<p className="contact-title">Email</p>
-
 					<a
 						href="mailto:hugo970217@gmail.com"
 						className="contact-link"
@@ -26,47 +21,30 @@ const ContactsList: React.FC = () => {
 						hugo970217@gmail.com
 					</a>
 				</div>
-
 			</li>
 
 			<li className="contact-item">
-
-				<div className="icon-box">
-					<TbPhoneCalling />
-				</div>
-
+				<IconBox icon={TbPhoneCalling} />
 				<div className="contact-info">
 					<p className="contact-title">Phone</p>
-
 					<p className="contact-link">404 Not Found 📲</p>
 				</div>
-
 			</li>
 
 			<li className="contact-item">
-
-				<div className="icon-box">
-					<IoCalendarOutline />
-				</div>
-
+				<IconBox icon={IoCalendarOutline} />
 				<div className="contact-info">
 					<p className="contact-title">Birthday</p>
 					<time dateTime="2002-01-27">Jan. 27, 2002 🐻</time>
 				</div>
-
 			</li>
 
 			<li className="contact-item">
-
-				<div className="icon-box">
-					<MdOutlineLocationOn />
-				</div>
-
+				<IconBox icon={MdOutlineLocationOn} />
 				<div className="contact-info">
 					<p className="contact-title">Location</p>
 					<address>Taipei, Taiwan 🇹🇼</address>
 				</div>
-
 			</li>
 		</ul>
 	);
