+++ src/styles/contact-list.css
@@ -15,30 +15,6 @@ time {
     gap: 16px;
 }
 
-.icon-box {
-    position: relative;
-    background: var(--border-gradient-onyx);
-    width: 30px;
-    height: 30px;
-    border-radius: 8px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 16px;
-    color: var(--orange-yellow-crayola);
-    box-shadow: var(--shadow-1);
-    z-index: 1;
-}
-
-.icon-box::before {
-    content: "";
-    position: absolute;
-    inset: 1px;
-    background: var(--eerie-black-1);
-    border-radius: inherit;
-    z-index: -1;
-}
-
 .contact-info {
     max-width: calc(100% - 46px);
     width: calc(100% - 46px);
@@ -63,13 +39,6 @@ time {
 @media (min-width: 580px) {
     .contacts-list {
         gap: 20px;
-    }
-
-    .icon-box {
-        width: 48px;
-        height: 48px;
-        border-radius: 12px;
-        font-size: 18px;
     }
 
     .contact-info {
